move fallback implementation from KDL to KinematicsBase
